% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_tx_names}
\alias{check_tx_names}
\title{Check validity of transcript vectors}
\usage{
check_tx_names(tx1, tx2, arg_name1, arg_name2)
}
\arguments{
\item{tx1}{A \code{character()} vector of GENCODE or ENSEMBL transcripts.}

\item{tx2}{A \code{character()} vector of GENCODE or ENSEMBL transcripts.}

\item{arg_name1}{A \code{character(1)} vector of description of tx1}

\item{arg_name2}{A \code{character(1)} vector of description of tx2}
}
\value{
A
\code{character()} vector of transcripts in \code{tx2} that are in \code{tx1}.
}
\description{
This function is used to check if the tx1 and tx2 are GENCODE or ENSEMBL and print an error message if it's not and return a character vector of transcripts in tx2 that are in tx1.
}
\examples{
sig_transcripts = select_transcripts("cell_component")
check_tx_names(rownames(covComb_tx_deg), sig_transcripts, 'rownames(covComb_tx_deg)', 'sig_transcripts')
}
