% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covComb_tx_deg-data.R
\docType{data}
\name{covComb_tx_deg}
\alias{covComb_tx_deg}
\title{RSE object of RNA-seq data that serves as output for degradation analysis}
\format{
A \link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
}
\description{
This data was generated from an experiment using degraded RNA-seq samples
post-mortem brain tissue. The transcripts included are the result of the qsva
expanded framework study and will be used to remove the effect of degradation
in bulk RNA-seq data.
}
\seealso{
\link{getPCs} \link{k_qsvs} \link{getDegTx}
}
\keyword{datasets}
