% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEqual.R
\name{DEqual}
\alias{DEqual}
\title{DEqual plot}
\usage{
DEqual(DE)
}
\arguments{
\item{DE}{a \code{data.frame()} with one column containing the t-statistics from
Differential Expression, typically generated with \code{limma::topTable()}.
The \code{rownames(DE)} should be transcript GENCODE IDs.}
}
\value{
a \code{ggplot} object of the DE t-statistic vs the DE statistic from degradation
}
\description{
TODO description. Maybe include links to published versions of this plot.
}
\examples{

## Random differential expression t-statistics for the same transcripts
## we have degradation t-statistics for in `degradation_tstats`.
set.seed(101)
random_de <- data.frame(
    t = rt(nrow(degradation_tstats), 5),
    row.names = sample(rownames(degradation_tstats), nrow(degradation_tstats))
)

## Create the DEqual plot
DEqual(random_de)
}
