% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBonfTx.R
\name{getBonfTx}
\alias{getBonfTx}
\title{PCs from transcripts}
\usage{
getBonfTx(covComb_tx)
}
\arguments{
\item{covComb_tx}{Ranged Summarizeed Experiment with only trancsripts selected for qsva}
}
\value{
prcomp object generated by taking the pcs of degraded transcripts
}
\description{
This function returns the pcs from the obtained RangedSummarizedExperiment object of selected transcripts
}
\examples{
getBonfTx(covComb_tx_deg)
 ## TODO
}
