% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPCs.R
\name{getPCs}
\alias{getPCs}
\title{PCs from features}
\usage{
getPCs(rse, assayname = "tpm")
}
\arguments{
\item{rse}{Ranged Summarizeed Experiment with only features selected for qsva}

\item{assayname}{character string specifying the name of the assay desired in rse}
}
\value{
prcomp object generated by taking the pcs of degraded features
}
\description{
This function returns the pcs from the obtained RangedSummarizedExperiment object of selected features
}
\examples{
getPCs(rse, "tpm")
}
