% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_tx_names}
\alias{normalize_tx_names}
\title{Remove version number from Gencode/Ensembl transcript names}
\usage{
normalize_tx_names(txnames)
}
\arguments{
\item{txnames}{A \code{character()} vector of GENCODE or ENSEMBL transcript IDs}
}
\value{
A
\code{character()} vector of transcript names without versioning
}
\description{
This function removes the Gencode/ENSEMBL version from the transcript ID, while protecting _PAR_Y suffixes if present
}
\examples{
ensIDs <- normalize_tx_names(rownames(rse_tx))
}
