% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rse_tx-data.R
\docType{data}
\name{rse_tx}
\alias{rse_tx}
\title{Example of RSE object with RNA-seq transcript quantification data}
\format{
A \link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
}
\description{
This data is a \link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
with transcript quantification data stored in an "tpm" assay. It is
used to demonstrate the use of qsvaR in bulk RNA-seq data.
}
\seealso{
\link{getPCs} \link{k_qsvs} \link{getDegTx} \link{qSVA}
}
\keyword{datasets}
