% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_transcripts.R
\name{select_transcripts}
\alias{select_transcripts}
\title{Select degradation related transcripts}
\usage{
select_transcripts(type)
}
\arguments{
\item{type}{A \code{character(1)} specifying the transcripts set type. These
were determined by Joshua M. Stolz et al, 2022.}
}
\value{
A \code{character()} with the transcript IDs.
}
\description{
TODO Description
}
\examples{
select_transcripts()
}
