% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcripts-data.R
\docType{data}
\name{transcripts}
\alias{transcripts}
\title{Transcripts for Degradation Models}
\format{
A \code{list()} with character strings containing the transcripts selected by each model.
Each string is a GENCODE transcript IDs.
}
\description{
Description TODO.
}
\seealso{
\link{select_transcripts}
}
\keyword{datasets}
