% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degradation_tstats-data.R
\docType{data}
\name{degradation_tstats}
\alias{degradation_tstats}
\title{Degradation time t-statistics}
\format{
A \code{data.frame()} with the \code{t} statistics for degradation time. The
\code{rownames()} are the GENCODE transcript IDs.
}
\description{
These t-statistics are derived from the same data that was used for
\link{rse_tx}. They are the results from main model where we determined
the relationship with degradation time adjusting for the brain region (so
parallel degradation effects across brain regions). They are used for
plotting in \code{DEqual()}.
}
\seealso{
\link{DEqual}
}
\keyword{datasets}
