% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEqual.R
\name{DEqual}
\alias{DEqual}
\title{Title}
\usage{
DEqual(DE)
}
\arguments{
\item{DE}{data.frame with one column of the t-statistic from Differential Expression. The row names are the transcripts names}
}
\value{
a ggplot object of the DE t-statistc vs the DE statistic from degradation
}
\description{
Title
}
\examples{
TODO
}
