% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEqual.R
\name{DEqual}
\alias{DEqual}
\title{Differential expression quality (DEqual) plot}
\usage{
DEqual(DE)
}
\arguments{
\item{DE}{a \code{data.frame()} with one column containing the t-statistics from
Differential Expression, typically generated with \code{limma::topTable()}.
The \code{rownames(DE)} should be transcript GENCODE IDs.}
}
\value{
a \code{ggplot} object of the DE t-statistic vs
the DE statistic from degradation
}
\description{
A DEqual plot compares the effect of RNA degradation from an independent
degradation experiment on the y axis to the effect of the outcome of
interest. They were orignally described by Jaffe et al, PNAS, 2017
\url{https://doi.org/10.1073/pnas.1617384114}. Other DEqual versions are
included in Collado-Torres et al, Neuron, 2019
\url{https://doi.org/10.1016/j.neuron.2019.05.013}. This function compares your
t-statistics of interest computed on transcripts against the
t-statistics from degradation time adjusting for the six brain regions from
degradation experiment data used for determining \code{covComb_tx_deg}.
}
\examples{

## Random differential expression t-statistics for the same transcripts
## we have degradation t-statistics for in `degradation_tstats`.
set.seed(101)
random_de <- data.frame(
    t = rt(nrow(degradation_tstats), 5),
    row.names = sample(
        rownames(degradation_tstats),
        nrow(degradation_tstats)
    )
)

## Create the DEqual plot
DEqual(random_de)
}
