% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qsvs.R
\name{get_qsvs}
\alias{get_qsvs}
\title{Generate matrix of qsvs}
\usage{
get_qsvs(qsvPCs, k)
}
\arguments{
\item{qsvPCs}{prcomp object generated by taking
the pcs of degraded transcripts}

\item{k}{number of qsvs to be included.}
}
\value{
matrix with k principal components for each sample.
}
\description{
Using the pcs and the k number of components be included,
we generate the qsva matrix.
}
\examples{
qsv <- getPCs(rse_tx, "tpm")
get_qsvs(qsv, 2)
}
