% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDegTx.R
\name{getDegTx}
\alias{getDegTx}
\title{Obtain expression matrix for degraded transcripts}
\usage{
getDegTx(
  rse_tx,
  type = "cell_component",
  sig_transcripts = select_transcripts(type)
)
}
\arguments{
\item{rse_tx}{A \link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class} object containing
the transcript data desired to be studied.}

\item{type}{A \code{character(1)} specifying the transcripts set type. These
were determined by Joshua M. Stolz et al, 2022.}

\item{sig_transcripts}{A list of transcripts determined to have degradation signal in the qsva expanded paper.}
}
\value{
A
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
object.
}
\description{
This function is used to obtain a \link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class} of transcripts and their expression values #' These transcripts are selected based on a prior study of RNA degradation in postmortem brain tissues. This object can later be used to obtain the principle components
necessary to remove the effect of degradation in differential expression.
}
\examples{
getDegTx(covComb_tx_deg)
}
