% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDegTx.R
\name{getDegTx}
\alias{getDegTx}
\title{Obtain expression matrix for degraded transcripts}
\usage{
getDegTx(
  rse_tx,
  type = c("cell_component", "standard", "top1500"),
  sig_transcripts = NULL,
  assayname = "tpm"
)
}
\arguments{
\item{rse_tx}{A \link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
object containing the transcript data desired to be studied.}

\item{type}{A \code{character(1)} specifying the transcripts set type.
These were determined by Joshua M. Stolz et al, 2022. Here the names "cell_component", "top1500",
and "standard" refer to models that were determined to be effective in removing degradation effects.
The "standard" model involves taking the union of the top 1000 transcripts
associated with degradation from the interaction model and the main effect model.
The "top1500" model is the same as the "standard model except the
union of the top 1500 genes associated with degradation is selected.
The most effective of our models, "cell_component", involved deconvolution of
the degradation matrix to determine the proportion of cell types within our studied tissue.
These proportions were then added to our \code{model.matrix()} and the union of the top 1000 transcripts in the interaction model,
the main effect model, and the cell proportions model were used to generate this model of qSVs.}

\item{sig_transcripts}{A list of transcripts determined to have degradation signal in the qsva expanded paper.}

\item{assayname}{character string specifying the name of the assay desired in rse_tx}
}
\value{
A
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
object.
}
\description{
This function is used to obtain a \link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
of transcripts and their expression values #' These transcripts are selected based on a prior study of RNA degradation in
postmortem brain tissues. This object can later be used to obtain the principle components
necessary to remove the effect of degradation in differential expression.
}
\examples{
degTx <- getDegTx(rse_tx, "standard")
}
