% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPCs.R
\name{getPCs}
\alias{getPCs}
\title{PCs from transcripts}
\usage{
getPCs(rse_tx, assayname = "tpm")
}
\arguments{
\item{rse_tx}{Ranged Summarizeed Experiment with only trancsripts selected for qsva}

\item{assayname}{character string specifying the name of the assay desired in rse_tx}
}
\value{
prcomp object generated by taking the pcs of degraded transcripts
}
\description{
This function returns the pcs from the obtained RangedSummarizedExperiment object of selected transcripts
}
\examples{
getPCs(covComb_tx_deg, "tpm")
}
