% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_qsvs.R
\name{k_qsvs}
\alias{k_qsvs}
\title{Apply num.sv algorithm to determine the number of pcs to be included}
\usage{
k_qsvs(rse_tx, mod, assayname)
}
\arguments{
\item{rse_tx}{A \link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class} object containing
the transcript data desired to be studied.}

\item{mod}{Model Matrix with necessary variables the you would
model for in differential expression}

\item{assayname}{character string specifying the name of the assay desired in rse_tx}
}
\value{
integer representing number of pcs to be included
}
\description{
Apply num.sv algorithm to determine the number of pcs to be included
}
\examples{
mod <- model.matrix(~ mitoRate + Region + rRNA_rate + totalAssignedGene + RIN,
    data = colData(covComb_tx_deg)
)
k_qsvs(covComb_tx_deg, mod, "tpm")

}
