% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_qsvs.R
\name{k_qsvs}
\alias{k_qsvs}
\title{Apply num.sv algorithm to determine the number of pcs to be included}
\usage{
k_qsvs(covComb_tx, mod_tx)
}
\arguments{
\item{covComb_tx}{Ranged Summarizeed Experiment with only trancsripts selected for qsva}

\item{mod_tx}{Model Matrix with necessary variables the you would model for in differential expression}
}
\value{
integer representing number of pcs to be included
}
\description{
Apply num.sv algorithm to determine the number of pcs to be included
}
\examples{
mod_tx <- model.matrix(~ mitoRate + Region + rRNA_rate + totalAssignedGene + RIN , data = colData(covComb_tx_deg))
k_qsvs(covComb_tx_deg,mod_tx)
}
