% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qSVA.R
\name{qSVA}
\alias{qSVA}
\title{A wrapper function used to perform qSVA in one step.}
\usage{
qSVA(
  rse,
  type = c("cell_component", "top1000", "top1500"),
  sig_features = NULL,
  mod,
  assayname
)
}
\arguments{
\item{rse}{A \link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
object containing the feature data desired to be studied.}

\item{type}{a character string specifying which model you would
like to use from the sets of signature features identified
by the qsvaR package. This can be omitted if a custom set of
features is provided to sig_features.}

\item{sig_features}{A list of feature IDs that are associated
with degradation signal. Specifying a \code{character()} input with ENSEMBL
feature IDs (whose values should match entries in \code{rownames(rse)}).
This argument provides a custom list of features for adjusting
for degradation; this should be used instead of the \code{type} argument.}

\item{mod}{Model Matrix with necessary variables the you would
model for in differential expression}

\item{assayname}{character string specifying the name of
the assay desired in rse}
}
\value{
matrix with k principal components for each sample
}
\description{
A wrapper function used to perform qSVA in one step.
}
\examples{
## First we need to define a statistical model. We'll use the example
## rse_tx data. Note that the model you'll use in your own data
## might look different from this model.
mod <- model.matrix(~ mitoRate + Region + rRNA_rate + totalAssignedGene + RIN,
    data = colData(rse)
)

## To ensure that the results are reproducible, you will need to set a
## random seed with the set.seed() function. Internally, we are using
## sva::num.sv() which needs a random seed to ensure reproducibility of the
## results.
set.seed(20230621)
qSVA(rse = rse, type = "cell_component", mod = mod, assayname = "tpm")

}
